requirements.txt
facenet_pytorch==2.5.3
Flask==2.2.2
matplotlib==3.7.0
matplotlib==3.8.4
numpy==1.23.5
numpy==1.22.3
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
Pillow==9.5.0
Pillow==9.4.0
Pillow==10.3.0
scikit_learn==1.2.1
seaborn==0.13.2
torch==2.0.1
torchvision==0.15.2
